backend/app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
ajsonrpc==1.2.0
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
apturl==0.5.2
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.29.0
<<<<<<< Updated upstream
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
=======
attrs==25.3.0
Babel==2.8.0
backcall==0.2.0
bcc==0.18.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
beniget==0.4.1
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
bottle==0.13.2
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.5.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.0.6
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dataclasses-json==0.6.7
dbus-fast==2.44.1
dbus-python==1.2.18
debugpy==1.8.5
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.7.0
distro-info==1.1+ubuntu0.2
docopt==0.6.2
dronecan==1.0.26
duplicity==0.8.21
empy==3.3.4
evdev==1.6.1
exceptiongroup==1.3.0
fastapi==0.115.2
fasteners==0.14.1
fido2==0.9.1
filetype==1.2.0
flake8==7.2.0
Flask==2.0.2
Flask-Cors==3.0.10
fonttools==4.29.1
frozenlist==1.2.0
fs==2.4.12
future==1.0.0
gast==0.6.0
gcovr==5.0
geocoder==1.38.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.185.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.71.0
greenlet==3.1.1
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==21.2.0
gyp==0.1
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.2
httptools==0.7.1
>>>>>>> Stashed changes
httpx==0.27.2
httpx-sse==0.4.3
idna==3.10
importlib-metadata==4.6.4
iniconfig==2.3.0
intelhex==2.3.0
ipykernel==6.29.5
itsdangerous==2.0.1
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.0.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
junitparser==3.2.0
jupyter_client==8.6.3
jupyterlab_pygments==0.3.0
keyring==23.5.0
kiwisolver==1.3.2
langchain==0.3.27
langchain-community==0.3.25
langchain-core==0.3.79
langchain-google-genai==2.0.7
langchain-text-splitters==0.3.11
langsmith==0.3.45
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lockfile==0.12.2
louis==3.20.0
lxml==5.4.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
marshmallow==3.25.1
matplotlib==3.5.1
matplotlib-inline==0.1.7
MAVProxy==1.8.71
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
multidict==5.1.0
mypy_extensions==1.1.0
narwhals==2.5.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.10.6
Nuitka==2.7.14
numpy==2.2.6
oauthlib==3.2.0
olefile==0.46
ordered-set==4.1.0
orjson==3.11.4
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
paramiko==2.9.3
patchify==0.2.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.0.1
pipreqs==0.5.0
platformio==6.1.16
pluggy==1.6.0
ply==3.11
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
proton-core==0.7.3
proton-keyring-linux==0.2.1
proton-keyring-linux-secretservice==0.1.0
proton-vpn-api-core==0.48.0
proton-vpn-daemon==0.13.4
proton-vpn-gtk-app==4.12.0
proton-vpn-killswitch-network-manager==0.6.1
proton-vpn-killswitch-network-manager-wireguard==0.2.1
proton-vpn-lib==0.1.1
proton-vpn-logger==0.2.1
proton-vpn-network-manager==0.13.4
proton-vpn-network-manager-openvpn==0.1.1
proton-vpn-network-manager-wireguard==0.4.7
psutil==6.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
pycodestyle==2.13.0
pycups==2.0.1
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
pyelftools==0.31
pyflakes==3.3.2
pygame==2.6.1
Pygments==2.11.2
PyGObject==3.42.1
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
PyJWT==2.3.0
pymacaroons==0.13.0
pymavlink==2.4.41
PyNaCl==1.5.0
pynmeagps==1.0.43
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
pytest==8.3.3
pytest-asyncio==0.24.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-magic==0.4.24
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.2.0
ratelim==0.1.6
redis==4.2.2
referencing==0.36.2
reportlab==3.6.8
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.27.1
RPi.GPIO==0.7.1
rsa==4.9
scipy==1.8.0
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
semantic-version==2.10.0
sentry-sdk==1.4.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
socketio==0.2.1
soupsieve==2.6
SQLAlchemy==2.0.36
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.39.2
streamlit==1.50.0
sympy==1.9
systemd-python==234
tabulate==0.9.0
tenacity==9.0.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tinycss2==1.4.0
toml==0.10.2
tomli==2.3.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==4.2.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.30.6
uvloop==0.22.1
wadllib==1.3.6
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.0
Werkzeug==2.0.2
wrapt==1.17.3
wsproto==1.2.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
yarg==0.1.9
yarl==1.22.0
zipp==1.0.0
zstandard==0.23.0
